Account for PCC having a non-zero base
This isn't quite right, since in theory the base could be
(Reloc Base + Function Base), and we only want Function Base to be
subtracted. In practice, Reloc Base will be C0's base, so we could use
that to correct it, but for my purposes that will always be 0 and is
therefore ignored. In future everything should be using capabilities
anyway, so this will go away.