[WIP, RISCV] Lower inline asm constraint A for RISC-V
Summary:
This allows arguments with the constraint A to be lowered to input nodes for RISC-V, which implies a memory address stored in a register.

This patch adds the minimal amount of code required to get operands with the right constraints to compile. Semantically it would be nice to be able to treat the constraint as a `C_RegisterClass` and then verify and lower the register as a memory operand.

Reviewers: asb

Subscribers: eraman, rbar, johnrusso, simoncook, apazos, sabuasal, niosHD, kito-cheng, shiva0217, jrtc27, zzheng, edward-jones, mgrang, rogfer01, MartinMosbeck, brucehoult, the_o, rkruppe, PkmX, jocewei, llvm-commits

Differential Revision: https://reviews.llvm.org/D54296