[RISCV] Add codegen support for 64-bit atomic load/store and atomicrmw
Summary:
I don't add support for cmpxchg, as D48131 is awaiting review and I'd rather not add it to the dependency chain.

New variants of the masked atomics intrinsics are introduced to support part-word atomicrmw are defined with i64 arguments and return values (as i64 is the only legal type for RV64).

Reviewers: jyknight, sabuasal, apazos

Subscribers: TheWaWaR, jrtc27, xxuejie, rbar, johnrusso, simoncook, niosHD, kito-cheng, shiva0217, zzheng, edward-jones, mgrang, rogfer01, MartinMosbeck, brucehoult, the_o, rkruppe, jfb, PkmX, jocewei, llvm-commits

Differential Revision: https://reviews.llvm.org/D53233