[RISCV] Attach VK_RISCV_CALL to symbols upon creation
Summary:
This patch replaces the addition of `VK_RISCV_CALL` in `RISCVMCCodeEmitter` by creating the `RISCVMCExpr` when tail/call are parsed, or in the codegen case when the callee symbols are created.

This required adding a new `CallSymbol` operand to allow only adding `VK_RISCV_CALL` to tail/call instructions.

This patch will allow further expansion of parsing and codegen to easily include PLT symbols which must generate the `R_RISCV_CALL_PLT` relocation.

Reviewers: asb, rogfer01, jrtc27

Subscribers: rbar, johnrusso, simoncook, apazos, sabuasal, niosHD, kito-cheng, shiva0217, zzheng, edward-jones, mgrang, MartinMosbeck, brucehoult, the_o, rkruppe, PkmX, jocewei, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D55560